# petshop-api/src/main/resources/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Pet Shop API",
    "description": "Pet shop inventory. Every endpoint requires an OIDC access token.",
    "version": "v1"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
  "tags": [
    {
      "name": "Pets",
      "description": "Pet shop inventory"
    }
  ],
  "paths": {
    "/api/pets/{id}": {
      "get": {
        "tags": [
          "Pets"
        ],
        "operationId": "get",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PetResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "oidc": []
          }
        ]
      },
      "put": {
        "tags": [
          "Pets"
        ],
        "operationId": "replace",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PetRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PetResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "oidc": []
          }
        ]
      },
      "delete": {
        "tags": [
          "Pets"
        ],
        "operationId": "delete",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "oidc": []
          }
        ]
      }
    },
    "/api/pets": {
      "get": {
        "tags": [
          "Pets"
        ],
        "operationId": "list",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "AVAILABLE",
                "PENDING",
                "SOLD"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PetResponse"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "oidc": []
          }
        ]
      },
      "post": {
        "tags": [
          "Pets"
        ],
        "operationId": "create",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PetRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PetResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "oidc": []
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
      "PetRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 100,
            "minLength": 0
          },
          "category": {
            "type": "string",
            "maxLength": 100,
            "minLength": 0
          },
          "status": {
            "type": "string",
            "enum": [
              "AVAILABLE",
              "PENDING",
              "SOLD"
            ]
          },
          "price": {
            "type": "number",
            "minimum": 0.0
          }
        },
        "required": [
          "category",
          "name",
          "price",
          "status"
        ]
      },
      "PetResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "category": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "AVAILABLE",
              "PENDING",
              "SOLD"
            ]
          },
          "price": {
            "type": "number"
          }
        }
      }
    },
    "securitySchemes": {
      "oidc": {
        "type": "oauth2",
        "flows": {
          "authorizationCode": {
            "authorizationUrl": "http://localhost:8081/realms/petshop/protocol/openid-connect/auth",
            "tokenUrl": "http://localhost:8081/realms/petshop/protocol/openid-connect/token",
            "scopes": {
              "pets:read": "Read the pet inventory",
              "pets:write": "Create, update and delete pets"
            }
          }
        }
      }
    }
  }
}